chore: relocate ramps-specific code
